requirements.txt
boto3==1.28.50
Django==4.2.5
django-ckeditor==6.7.0
django-cors-headers==4.2.0
django-storages==1.14
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
gunicorn==21.2.0
Markdown==3.4.4
num2words==0.5.12
phonenumbers==8.13.20
Pillow==10.0.1
psycopg2-binary==2.9.7
pyotp==2.9.0
python-decouple==3.8
razorpay==1.4.1
six==1.16.0
Unipath==1.1
